Avoid work if signed assembly already exists.
If the signed assembly folder already contains the output assembly and its Mvid matches the Mvid of the current assembly, avoid all the work and just use the existing output assembly.

This helps with performance and also build incrementality. If the output assembly is not touched unnecessarily it's timestamp won't change and it won't invalidate the rest of the build as it currently does.